backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.31
python-jose[cryptography]>=3.3.0
python-multipart==0.0.9
openai>=1.35.0
semantic-kernel>=1.0.0
python-dotenv==1.0.1
pydantic==2.8.2
pydantic-settings==2.3.4
aiofiles==24.1.0
bcrypt==4.1.3
email-validator>=2.0.0
